Event.go not EventJson
