+++ Makefile
@@ -1,2 +1,5 @@
 all:
+	go build -o bin/main *.go && ./bin/main
+
+i:
 	go build -o bin/main *.go && ./bin/main -i
